logmsg: move lib/type-hinting.{c,h} to lib/logmsg along with tests
