+++ lib/Makefile.am
@@ -175,7 +175,6 @@ pkginclude_HEADERS			+= \
 	lib/tls-support.h		\
 	lib/thread-utils.h		\
 	lib/tlscontext.h  		\
-	lib/type-hinting.h		\
 	lib/uuid.h			\
 	lib/userdb.h			\
 	lib/utf8utils.h			\
@@ -255,7 +254,6 @@ lib_libsyslog_ng_la_SOURCES		= \
 	lib/str-utils.c			\
 	lib/syslog-names.c		\
 	lib/string-list.c		\
-	lib/type-hinting.c		\
 	lib/ringbuffer.c		\
 	lib/crypto.c			\
 	lib/tlscontext.c		\
