+++ lib/logmsg/Makefile.am
@@ -15,20 +15,22 @@ logmsginclude_HEADERS =     \
  lib/logmsg/nvtable-serialize-endianutils.h \
  lib/logmsg/tags-serialize.h                \
  lib/logmsg/timestamp-serialize.h           \
- lib/logmsg/tags.h
+ lib/logmsg/tags.h			    \
+ lib/logmsg/type-hinting.h
 
-logmsg_sources =             \
- lib/logmsg/gsockaddr-serialize.c \
- lib/logmsg/logmsg.c              \
- lib/logmsg/logmsg-serialize.c    \
- lib/logmsg/logmsg-serialize-fixup.c \
- lib/logmsg/nvhandle-descriptors.c  \
- lib/logmsg/nvtable.c             \
- lib/logmsg/nvtable-serialize.c   \
+logmsg_sources =                       \
+ lib/logmsg/gsockaddr-serialize.c      \
+ lib/logmsg/logmsg.c                   \
+ lib/logmsg/logmsg-serialize.c         \
+ lib/logmsg/logmsg-serialize-fixup.c   \
+ lib/logmsg/nvhandle-descriptors.c     \
+ lib/logmsg/nvtable.c                  \
+ lib/logmsg/nvtable-serialize.c        \
  lib/logmsg/nvtable-serialize-legacy.c \
- lib/logmsg/tags-serialize.c      \
- lib/logmsg/timestamp-serialize.c \
- lib/logmsg/tags.c
+ lib/logmsg/tags-serialize.c           \
+ lib/logmsg/timestamp-serialize.c      \
+ lib/logmsg/tags.c		       \
+ lib/logmsg/type-hinting.c
 
 
 include lib/logmsg/tests/Makefile.am
