+++ lib/logmsg/tests/CMakeLists.txt
@@ -6,3 +6,4 @@ add_unit_test(CRITERION TARGET test_gsockaddr_serialize)
 add_unit_test(CRITERION LIBTEST TARGET test_log_message)
 add_unit_test(CRITERION TARGET test_logmsg_ack)
 add_unit_test(CRITERION TARGET test_nvhandle_desc_array)
+add_unit_test(CRITERION TARGET test_type_hints)
