+++ lib/logmsg/type-hinting.c
