+++ lib/logmsg/type-hinting.h
