+++ lib/tests/CMakeLists.txt
@@ -1,6 +1,5 @@
 add_unit_test(CRITERION TARGET test_cfg_lexer_subst)
 add_unit_test(CRITERION TARGET test_cfg_tree)
-add_unit_test(CRITERION TARGET test_type_hints)
 add_unit_test(CRITERION TARGET test_parse_number)
 add_unit_test(CRITERION TARGET test_reloc)
 add_unit_test(CRITERION TARGET test_hostname)
