+++ lib/tests/Makefile.am
@@ -2,7 +2,6 @@ lib_tests_TESTS		= \
 	lib/tests/test_cfg_lexer_subst	\
 	lib/tests/test_lexer_block	\
 	lib/tests/test_cfg_tree		\
-	lib/tests/test_type_hints	\
 	lib/tests/test_parse_number	\
 	lib/tests/test_reloc		\
 	lib/tests/test_hostname		\
@@ -53,10 +52,6 @@ lib_tests_test_cfg_tree_CFLAGS		=	\
 lib_tests_test_cfg_tree_LDADD		=	\
 	$(TEST_LDADD)
 
-lib_tests_test_type_hints_CFLAGS	=	\
-	$(TEST_CFLAGS)
-lib_tests_test_type_hints_LDADD	=	\
-	$(TEST_LDADD)
 
 lib_tests_test_parse_number_CFLAGS	=	\
 	$(TEST_CFLAGS)
