+++ lib/value-pairs/value-pairs.c
@@ -25,9 +25,9 @@
 #include "value-pairs/internals.h"
 #include "value-pairs/value-pairs.h"
 #include "logmsg/logmsg.h"
+#include "logmsg/type-hinting.h"
 #include "template/templates.h"
 #include "template/macros.h"
-#include "type-hinting.h"
 #include "cfg-parser.h"
 #include "string-list.h"
 #include "scratch-buffers.h"
